:recycle: 메인 페이지 템플릿 컴포넌트 분리
