+++ src/Pages/Main/Aside/index.tsx
@@ -1,0 +1,11 @@
+import { SideTab } from '@/components/server/Filter/SideTab';
+
+import * as S from './style';
+
+export function Aside() {
+  return (
+    <S.StyledAside>
+      <SideTab />
+    </S.StyledAside>
+  );
+}
