+++ src/Pages/Main/Aside/style.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const StyledAside = styled.aside`
+  position: fixed;
+  width: 15%;
+  left: 5%;
+  top: 30%;
+`;
