+++ src/Pages/Main/Contents/style.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const MainContents = styled.section`
+  display: flex;
+  width: 60%;
+  margin: 0 auto;
+  flex-wrap: wrap;
+  min-height: 100vh;
+`;
